Change `m` alias to a custom function
The `m` gem doesn't work well with Rails when parallel tests are
enabled. When `m`'s argument is a file path, a duplicate `at_exit` hook
is registered causing the parallel tests to fork again after all the
tests have run. This causes an error because the DRb server is no longer
running at this point.

Write a shell function that looks for `config/environment.rb` to
determine if the current project is a Rails app. If so, forward
arguments to `rails test`, which supports selecting tests using a line
number in the file path (i.e. `test/models/payment_test:6`).

If `config/environment.rb` does not exist, forward the arguments to
`bundle exec m`.

When in a Rails app (`config/environment.rb` exists), and the arguments
to our custom `m` is "test" or empty, use "test/*" as the argument so
all tests are run. Rails' default is to exclude system tests when
running `rails test`.

The custom `m` function is a convenience since I've learned to use `m`
and want to avoid context switching between `m` in non-Rails projects
and `rails test` (or an alias for it) in Rails projects.